API V1

+++ src/services/pokemon_service.ts
@@ -12,7 +12,9 @@ export class PokemonService {
         this.pokemonRepository = AppDataSource.getRepository(Pokemon);
     }
 
-    async createPokemon(pokemonData: PokemonInterface, req: Request): Promise<Pokemon> {
+    async createPokemon(req: Request): Promise<Pokemon> {
+        const pokemonData: PokemonInterface = req.body;
+
         const imagePath = req.file?.path;  
         pokemonData.imagePath = imagePath;
 
@@ -20,17 +22,24 @@ export class PokemonService {
         return await this.pokemonRepository.save(pokemon);
     }
 
-    async updatePokemon(id: number, pokemonData: Partial<PokemonInterface>, req: Request): Promise<Pokemon | null> {
+    async updatePokemon(id: number,req: Request): Promise<Pokemon | null> {
+
         const pokemon = await this.pokemonRepository.findOneBy({ id });
+
         if (pokemon) {
             const pokemonData: Partial<PokemonInterface> = req.body
 
             if(req.file?.path){
                 if(pokemon.imagePath){
-                    fs.unlinkSync(pokemon.imagePath)
+                    try {
+                        fs.unlinkSync(pokemon.imagePath)
+                    } catch (error) {
+                        console.error(`Erro ao deletar a imagem: ${error}`);
+                    }
                 }
             pokemonData.imagePath = req.file.path
-            }   
+            }
+
             const updatedPokemon = this.pokemonRepository.merge(pokemon, pokemonData);
             return await this.pokemonRepository.save(updatedPokemon);
         }
@@ -38,11 +47,27 @@ export class PokemonService {
     }
 
     async deletePokemon(id: number): Promise<boolean> {
-        const result = await this.pokemonRepository.delete(id);
-        return result.affected !== 0;
+        const pokemon = await this.pokemonRepository.findOneBy({id})
+
+        if (pokemon) {
+            if (pokemon.imagePath) {
+                try {
+                    fs.unlinkSync(pokemon.imagePath); 
+                } catch (error) {
+                    console.error(`Erro ao deletar a imagem: ${error}, \nVocê pode ignorar isso, essa mensagem só vai aparecer caso vc tenha deletado o pokemon com imagem, e deletado a imagem antes de apagar.`);
+                }
+            }
+            const result = await this.pokemonRepository.delete(id);
+            return result.affected !== 0;
+        }
+        return false;
     }
 
     async getAllPokemons(): Promise<Pokemon[]> {
         return await this.pokemonRepository.find();
+    }
+
+    async getPokemonById(id: number): Promise<Pokemon | null> {
+        return await this.pokemonRepository.findOneBy({ id });
     }
 }
